requirements.txt
streamlit
stripe
mysql-connector-python
scikit-learn
pandas